move request to utils
