+++ src/admin.ts
@@ -2,7 +2,7 @@ import { Menus } from './resources/project/admin'
 
 import type { Blutui } from './blutui'
 import type { Project } from './project'
-import { Request } from './request'
+import { Request } from './utils/request'
 
 export class Admin extends Request {
   readonly menus = new Menus(this)
